# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Twin Fisiologico API",
    "description": "Stateless physiology analytics API. HTTP routers are thin; application services orchestrate engines under `engines/`.",
    "version": "5.1.0",
    "x-codegen": {
      "typescript_output": "frontend/src/api/generated/schema.ts",
      "regenerate": "make openapi-frontend"
    }
  },
  "paths": {
...
    "/ride/ingest": {
      "post": {
        "tags": [
          "ride"
        ],
        "summary": "Ingest ride FIT and update power curve",
        "description": "Parse one ride FIT, merge into the rolling MMP curve and return persistable state.",
        "operationId": "rideIngest",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_rideIngest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated rolling power curve",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideIngestResponse",
                  "properties": {
                    "curve": {
                      "additionalProperties": true,
                      "type": "object",
                      "title": "Curve",
                      "description": "Rolling power-duration curve to persist."
                    },
                    "mmp_for_profiler": {
                      "additionalProperties": {
                        "type": "number"
                      },
                      "type": "object",
                      "title": "Mmp For Profiler",
                      "description": "MMP slice for metabolic profiler."
                    },
                    "improvements": {
                      "type": "integer",
                      "title": "Improvements",
                      "description": "Number of duration records improved on this ride."
                    },
                    "ride_usable": {
                      "type": "boolean",
                      "title": "Ride Usable",
                      "description": "Whether the ride should contribute to the curve."
                    },
                    "profile_should_refresh": {
                      "type": "boolean",
                      "title": "Profile Should Refresh",
                      "description": "Whether anchor/snapshot should be recomputed."
                    },
                    "notes": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array",
                      "title": "Notes"
                    }
                  },
                  "type": "object",
                  "required": [
                    "curve",
                    "mmp_for_profiler",
                    "improvements",
                    "ride_usable",
                    "profile_should_refresh"
                  ],
                  "title": "RideIngestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable FIT or activity",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ride/update-profile": {
      "post": {
        "tags": [
          "ride"
        ],
        "summary": "Update profile from ride MMP",
        "description": "Bayesian-style profile update using stored anchor and ride MMP.",
        "operationId": "rideUpdateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ride/summary": {
      "post": {
        "tags": [
          "ride"
        ],
        "summary": "Full activity workout summary",
        "description": "Orchestrated per-activity report (power, zones, cardiac, HRV, mader_durability).",
        "operationId": "rideSummary",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_rideSummary"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ride/durability": {
      "post": {
        "tags": [
          "ride"
        ],
        "summary": "Mader session durability",
        "description": "Mechanistic CP residual curve and sustainable power targets for the ride.",
        "operationId": "rideDurability",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_rideDurability"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable FIT or activity",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/team/calibration/update": {
      "post": {
        "tags": [
          "team"
        ],
        "summary": "Update team calibration model",
        "description": "Add validated test events. Requires pre-test predicted_value per event.",
        "operationId": "teamCalibrationUpdate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCalibrationUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/team/calibration/apply": {
      "post": {
        "tags": [
          "team"
        ],
        "summary": "Apply team calibration",
        "description": "Apply learned correction to a snapshot or single parameter estimate.",
        "operationId": "teamCalibrationApply",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCalibrationApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_rideDurability": {
        "properties": {
          "weight_kg": {
            "type": "number",
            "title": "Weight Kg"
          },
          "metabolic_snapshot_json": {
            "type": "string",
            "title": "Metabolic Snapshot Json",
            "description": "Required successful metabolic snapshot."
          },
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "power_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power Json"
          }
        },
        "type": "object",
        "required": [
          "weight_kg",
          "metabolic_snapshot_json"
        ],
        "title": "Body_rideDurability"
      },
      "Body_rideIngest": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Ride FIT file."
          },
          "ride_date": {
            "type": "string",
            "title": "Ride Date",
            "description": "ISO date YYYY-MM-DD."
          },
          "weight_kg": {
            "type": "number",
            "title": "Weight Kg",
            "description": "Athlete weight in kg.",
            "default": 70.0
          },
          "stored_curve_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stored Curve Json",
            "description": "Previously persisted curve JSON (optional)."
          }
        },
        "type": "object",
        "required": [
          "file",
          "ride_date"
        ],
        "title": "Body_rideIngest"
      },
      "Body_rideSummary": {
        "properties": {
          "weight_kg": {
            "type": "number",
            "title": "Weight Kg",
            "description": "Athlete weight kg."
          },
          "ftp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ftp",
            "description": "Optional FTP override for zones."
          },
          "lthr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lthr",
            "description": "Optional LTHR override."
          },
          "gender": {
            "type": "string",
            "title": "Gender",
            "default": "MALE"
          },
          "training_years": {
            "type": "number",
            "title": "Training Years",
            "default": 10
          },
          "discipline": {
            "type": "string",
            "title": "Discipline",
            "default": "ENDURANCE"
          },
          "metabolic_snapshot_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metabolic Snapshot Json",
            "description": "Successful /profile/snapshot JSON to unlock durability sections."
          },
          "hrv_step_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hrv Step Seconds"
          },
          "hrv_max_windows": {
            "type": "integer",
            "title": "Hrv Max Windows",
            "default": 500
          },
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File",
            "description": "Ride FIT file."
          },
          "power_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power Json",
            "description": "Alternative: 1 Hz power JSON array."
          }
        },
        "type": "object",
        "required": [
          "weight_kg"
        ],
        "title": "Body_rideSummary"
      },
...
      "EnginePayload": {
        "properties": {},
        "additionalProperties": true,
        "type": "object",
        "title": "EnginePayload",
        "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RideIngestResponse": {
        "properties": {
          "curve": {
            "additionalProperties": true,
            "type": "object",
            "title": "Curve",
            "description": "Rolling power-duration curve to persist."
          },
          "mmp_for_profiler": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Mmp For Profiler",
            "description": "MMP slice for metabolic profiler."
          },
          "improvements": {
            "type": "integer",
            "title": "Improvements",
            "description": "Number of duration records improved on this ride."
          },
          "ride_usable": {
            "type": "boolean",
            "title": "Ride Usable",
            "description": "Whether the ride should contribute to the curve."
          },
          "profile_should_refresh": {
            "type": "boolean",
            "title": "Profile Should Refresh",
            "description": "Whether anchor/snapshot should be recomputed."
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "curve",
          "mmp_for_profiler",
          "improvements",
          "ride_usable",
          "profile_should_refresh"
        ],
        "title": "RideIngestResponse"
      },
...
      "TeamCalibrationApplyRequest": {
        "properties": {
          "calibration_model": {
            "additionalProperties": true,
            "type": "object",
            "title": "Calibration Model"
          },
          "parameter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameter",
            "description": "mlss, vo2max, vlamax, fatmax, map."
          },
          "predicted_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicted Value"
          },
          "snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot",
            "description": "Full metabolic snapshot to calibrate in one call."
          },
          "athlete_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Athlete Id"
          },
          "phenotype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phenotype"
          },
          "data_depth_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Data Depth Score",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "calibration_model"
        ],
        "title": "TeamCalibrationApplyRequest"
      },
      "TeamCalibrationUpdateRequest": {
        "properties": {
          "team_id": {
            "type": "string",
            "title": "Team Id"
          },
          "calibration_model": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibration Model",
            "description": "Previously persisted model; omit to start fresh."
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Events",
            "description": "Validation events with pre-test predicted_value."
          }
        },
        "type": "object",
        "required": [
          "team_id"
        ],
        "title": "TeamCalibrationUpdateRequest"
      },
...
      "UpdateProfileRequest": {
        "properties": {
          "anchor": {
            "additionalProperties": true,
            "type": "object",
            "title": "Anchor",
            "description": "Persisted measured anchor from POST /test/confirm."
          },
          "ride_mmp": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Ride Mmp",
            "description": "Ride MMP map {duration_seconds: watts}."
          },
          "athlete": {
            "$ref": "#/components/schemas/AthleteParams"
          },
          "as_of": {
            "type": "string",
            "title": "As Of",
            "description": "ISO date of the ride used for the update."
          },
          "load_factor": {
            "type": "number",
            "title": "Load Factor",
            "description": "Optional load scaling for Bayesian update.",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "anchor",
          "ride_mmp",
          "athlete",
          "as_of"
        ],
        "title": "UpdateProfileRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "{baseUrl}",
      "description": "Configurable deployment",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:8000",
          "description": "Set via VITE_API_BASE_URL in the frontend"
        }
      }
    }
  ],
...
}